+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Verb Changelog
 ## master
-- Fix Lisp code tags being evaluated in comments.
+- Fixed Lisp code tags being evaluated in comments.
+- Rebound `verb-send-request-on-point-other-window` to <kbd>C-c C-r C-s</kbd> in Verb mode.
+- Added new command `verb-send-request-on-point-other-window-stay`, bound to <kbd>C-c C-r C-r</kbd> in Verb mode.
+- Added new command `verb-kill-all-response-buffers`, bound to <kbd>C-c C-r C-k</kbd> in Verb mode.
+- Added new command `verb-re-send-request`, bound to <kbd>C-c C-r C-f</kbd> in Verb response body mode.
+- Remove `verb-headers-to-string` function.
 
 ## **1.3.0** - 2020-01-17
 - Request headers can now be commented out using `#`.
@@ -9,8 +14,8 @@
   - URL is now the first argument.
   - Arguments are now separated by (escaped) newlines.
 - Added `verb-export-request-on-point-curl`, `verb-export-request-on-point-verb` and `verb-export-request-on-point-human` commands.
-- Bind <kbd>C-c C-r C-u</kbd> to `verb-export-request-on-point-curl` in Verb mode.
-- Added `verb-set-var` command, bind it to <kbd>C-c C-r C-v</kbd>.
+- Bound <kbd>C-c C-r C-u</kbd> to `verb-export-request-on-point-curl` in Verb mode.
+- Added `verb-set-var` command, bound to <kbd>C-c C-r C-v</kbd> in Verb mode.
 
 ## **1.2.0** - 2020-01-16
 - Request specifications can now be exported to `curl`.